Basic keyboard interactions in place.
But it's worth noting that they show how slow it is to swap between views - I suspect we need to be loading
the thumbnail first and then swapping to the full image only once ready (we have some large GeoTIFFs for example
where this is really notable). This would also help with the press space to zoom from grid.